requirements.txt
apispec==6.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
pycountry==24.6.1
python-dotenv==1.1.0
SQLAlchemy==2.0.39
typing_extensions==4.13.0
webargs==8.6.0
Werkzeug==3.1.3
